checkpoint

+++ mandel_shared.py
@@ -1,5 +1,8 @@
 import os
 import select
+# from mpmath import mp, mpf
+
+# mp.dps = 300
 
 
 def get_params(fd):
@@ -9,8 +12,8 @@ def get_params(fd):
 
 
 def convert_x2r(x, x_center, size, zoom):
-    return x_center + zoom * float(x - size / 2) / size
+    return x_center + zoom * (x - size / 2.0) / size
 
 
 def convert_y2i(y, y_center, size, zoom):
-    return y_center + zoom * float(y - size / 2) / size
+    return y_center + zoom * (y - size / 2.0) / size
